Add sudospawner + kerberos example
Prompt the user for his/her kerberos password at spawn time so that a
Kerberos cred cache can be established on disk and chowned to the target
user. The chown cannot be done by the unprivileged user that owns the
hub process, so we skip cred cache creation at login time and instead do
it at spawn time after sudo'ing to the the target user (effectively
making the chown a no-op).